ranking: set latest rank for newly appeared user

+++ modules/jb-module-ranking/src/main/java/ru/caramel/juniperbot/module/ranking/persistence/entity/Ranking.java
@@ -39,6 +39,6 @@ public class Ranking extends BaseEntity {
     private long exp;
 
     @Column
-    private int rank;
+    private long rank;
 
 }
